+++ fixtures/coverall/src/lib.rs
@@ -11,8 +11,8 @@ use once_cell::sync::Lazy;
 
 mod traits;
 pub use traits::{
-    ancestor_names, get_traits, make_rust_getters, test_getters, test_round_trip_through_foreign,
-    test_round_trip_through_rust, Getters, NodeTrait,
+    ancestor_names, get_string_util_traits, get_traits, make_rust_getters, test_getters,
+    test_round_trip_through_foreign, test_round_trip_through_rust, Getters, NodeTrait, StringUtil,
 };
 
 static NUM_ALIVE: Lazy<RwLock<u64>> = Lazy::new(|| RwLock::new(0));