+++ fixtures/proc-macro/tests/bindings/test_proc_macro.swift
@@ -22,9 +22,13 @@ let obj2 = Object()
 assert(obj.isOtherHeavy(other: obj2) == .uncertain)
 
 let traitImpl = obj.getTrait(inc: nil)
-assert(traitImpl.name() == "TraitImpl")
-assert(obj.getTrait(inc: traitImpl).name() == "TraitImpl")
-assert(getTraitNameByRef(t: traitImpl) == "TraitImpl")
+assert(traitImpl.concatStrings(a: "foo", b: "bar") == "foobar")
+assert(obj.getTrait(inc: traitImpl).concatStrings(a: "foo", b: "bar") == "foobar")
+assert(concatStringsByRef(t: traitImpl, a: "foo", b: "bar") == "foobar")
+
+let traitImpl2 = obj.getTraitWithForeign(inc: nil)
+assert(traitImpl2.name() == "RustTraitImpl")
+assert(obj.getTraitWithForeign(inc: traitImpl2).name() == "RustTraitImpl")
 
 assert(enumIdentity(value: .true) == .true)